Remove redundant BaseBigQueryConnectorSmokeTest
