--- plugin/trino-bigquery/src/test/java/io/trino/plugin/bigquery/BaseBigQueryConnectorSmokeTest.java
@@ -1,36 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.trino.plugin.bigquery;
-
-import io.trino.testing.BaseConnectorSmokeTest;
-import io.trino.testing.TestingConnectorBehavior;
-
-public abstract class BaseBigQueryConnectorSmokeTest
-        extends BaseConnectorSmokeTest
-{
-    @Override
-    protected boolean hasBehavior(TestingConnectorBehavior connectorBehavior)
-    {
-        return switch (connectorBehavior) {
-            case SUPPORTS_TRUNCATE -> true;
-            case SUPPORTS_CREATE_MATERIALIZED_VIEW,
-                 SUPPORTS_CREATE_VIEW,
-                 SUPPORTS_MERGE,
-                 SUPPORTS_RENAME_SCHEMA,
-                 SUPPORTS_RENAME_TABLE,
-                 SUPPORTS_UPDATE -> false;
-            default -> super.hasBehavior(connectorBehavior);
-        };
-    }
-}
